# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-weather-guard",
    "version": "0.1.0",
    "description": "Go/no-go weather decisions for plans — forecasts and verdicts from Open-Meteo/NWS, keyless and live\n\nEvery paid route returns the purchased artifact in the 200 response body. Pay in USDC on Base or Solana — your client picks the rail.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com",
      "url": "https://github.com/nirholas/x402-weather-guard"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4024",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Deployed instance",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:4024"
        }
      }
    }
  ],
...
  "paths": {
    "/forecast": {
      "get": {
        "operationId": "forecast",
        "summary": "Hourly forecast for a point and window, plus active alerts ($0.001)",
        "description": "Hourly forecast for a point and window, plus active alerts\n\nPaid route — $0.001 in USDC. Pay on Base (EVM) or Solana; the 402 challenge lists both rails.",
        "tags": [
          "paid"
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "description": "Latitude, -90…90.",
            "schema": {
              "type": "number"
            },
            "example": 47.6062
          },
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "description": "Longitude, -180…180.",
            "schema": {
              "type": "number"
            },
            "example": -122.3321
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "description": "Length of the window in hours, 1–168. Default 24.",
            "schema": {
              "type": "integer"
            },
            "example": 6
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "description": "ISO-8601 start of the window. Default: now.",
            "schema": {
              "type": "string"
            },
            "example": "2026-08-07T03:00:00Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Hour-by-hour temperature, feels-like, precipitation probability and amount, wind, gusts, cloud, visibility, and plain-English conditions, with any active NWS alerts",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64 JSON settlement receipt.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "source",
                    "location",
                    "window",
                    "units",
                    "hourly",
                    "alerts",
                    "retrievedAt"
                  ],
                  "properties": {
                    "source": {
                      "type": "object",
                      "properties": {
                        "forecast": {
                          "type": "string",
                          "enum": [
                            "open-meteo"
                          ]
                        },
                        "alerts": {
                          "type": "string",
                          "enum": [
                            "nws"
                          ]
                        }
                      }
                    },
                    "location": {
                      "type": "object",
                      "properties": {
                        "requested": {
                          "type": "object",
                          "properties": {
                            "latitude": {
                              "type": "number"
                            },
                            "longitude": {
                              "type": "number"
                            }
                          }
                        },
                        "resolved": {
                          "type": "object",
                          "description": "The grid point Open-Meteo actually used.",
                          "properties": {
                            "latitude": {
                              "type": "number"
                            },
                            "longitude": {
                              "type": "number"
                            },
                            "elevationM": {
                              "type": [
                                "number",
                                "null"
                              ]
                            }
                          }
                        },
                        "timezone": {
                          "type": "string",
                          "enum": [
                            "UTC"
                          ]
                        }
                      }
                    },
                    "window": {
                      "type": "object",
                      "properties": {
                        "start": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "end": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "hours": {
                          "type": "integer"
                        }
                      }
                    },
                    "units": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "hourly": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "time",
                          "conditions"
                        ],
                        "properties": {
                          "time": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Top of the hour, UTC."
                          },
                          "temperatureC": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "feelsLikeC": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "humidityPct": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "precipitationProbabilityPct": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "precipitationMm": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "windKph": {
                            "type": [
                              "number",
                              "null"
                            ],
                            "description": "Sustained wind at 10 m."
                          },
                          "gustKph": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "cloudCoverPct": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "visibilityM": {
                            "type": [
                              "number",
                              "null"
                            ]
                          },
                          "weatherCode": {
                            "type": [
                              "integer",
                              "null"
                            ],
                            "description": "WMO weather code."
                          },
                          "conditions": {
                            "type": "string",
                            "description": "The WMO code in plain English."
                          }
                        }
                      }
                    },
                    "alerts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "event": {
                            "type": "string"
                          },
                          "severity": {
                            "type": "string",
                            "description": "NWS severity — Extreme, Severe, Moderate, Minor, Unknown."
                          },
                          "urgency": {
                            "type": "string"
                          },
                          "certainty": {
                            "type": "string"
                          },
                          "onset": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "ends": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "headline": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "areaDesc": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        }
                      }
                    },
                    "alertStatus": {
                      "type": "string",
                      "description": "`ok`, `ok (no active alerts)`, or `unavailable: <reason>`."
                    },
                    "retrievedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                },
                "example": {
                  "source": {
                    "forecast": "open-meteo",
                    "alerts": "nws"
                  },
                  "location": {
                    "requested": {
                      "latitude": 47.6062,
                      "longitude": -122.3321
                    },
                    "resolved": {
                      "latitude": 47.595562,
                      "longitude": -122.32443,
                      "elevationM": 59
                    },
                    "timezone": "UTC"
                  },
                  "window": {
                    "start": "2026-08-07T03:00:00.000Z",
                    "end": "2026-08-07T09:00:00.000Z",
                    "hours": 6
                  },
                  "units": {
                    "temperatureC": "°C",
                    "precipitationMm": "mm",
                    "precipitationProbabilityPct": "%",
                    "windKph": "km/h",
                    "gustKph": "km/h",
                    "visibilityM": "m"
                  },
                  "hourly": [
                    {
                      "time": "2026-08-07T03:00:00.000Z",
                      "temperatureC": 23.2,
                      "feelsLikeC": 22.7,
                      "humidityPct": 50,
                      "precipitationProbabilityPct": 0,
                      "precipitationMm": 0,
                      "windKph": 8.3,
                      "gustKph": 18.7,
                      "cloudCoverPct": 0,
                      "visibilityM": 39500,
                      "weatherCode": 0,
                      "conditions": "clear sky"
                    },
                    {
                      "time": "2026-08-07T04:00:00.000Z",
                      "temperatureC": 22.1,
                      "feelsLikeC": 21.8,
                      "humidityPct": 54,
                      "precipitationProbabilityPct": 0,
                      "precipitationMm": 0,
                      "windKph": 7.6,
                      "gustKph": 16.2,
                      "cloudCoverPct": 3,
                      "visibilityM": 41200,
                      "weatherCode": 0,
                      "conditions": "clear sky"
                    },
                    {
                      "time": "2026-08-07T05:00:00.000Z",
                      "temperatureC": 21,
                      "feelsLikeC": 20.9,
                      "humidityPct": 58,
                      "precipitationProbabilityPct": 0,
                      "precipitationMm": 0,
                      "windKph": 6.9,
                      "gustKph": 14.4,
                      "cloudCoverPct": 11,
                      "visibilityM": 42800,
                      "weatherCode": 1,
                      "conditions": "mainly clear"
                    }
                  ],
                  "alerts": [
                    {
                      "event": "Heat Advisory",
                      "severity": "Moderate",
                      "urgency": "Expected",
                      "certainty": "Likely",
                      "onset": "2026-08-07T11:00:00-07:00",
                      "ends": "2026-08-08T21:00:00-07:00",
                      "headline": "Heat Advisory issued August 6 at 2:11PM PDT by NWS Seattle WA",
                      "areaDesc": "King, WA; Pierce, WA; Snohomish, WA"
                    }
                  ],
                  "alertStatus": "ok",
                  "retrievedAt": "2026-08-07T03:04:11.882Z"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required (or payment rejected). The body lists both accepted rails.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequired"
                }
              }
            }
          },
          "502": {
            "description": "Upstream data source failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description."
          }
        }
      },
      "PaymentRequirements": {
        "type": "object",
        "description": "One accepted payment rail, per the x402 specification.",
        "required": [
          "scheme",
          "network",
          "maxAmountRequired",
          "resource",
          "payTo",
          "asset"
        ],
        "properties": {
          "scheme": {
            "type": "string",
            "enum": [
              "exact"
            ]
          },
          "network": {
            "type": "string",
            "enum": [
              "base-sepolia",
              "base",
              "solana",
              "solana-devnet"
            ],
            "description": "EVM or Solana network this rail settles on."
          },
          "maxAmountRequired": {
            "type": "string",
            "description": "Price in USDC base units (6 decimals)."
          },
          "resource": {
            "type": "string",
            "format": "uri"
          },
          "description": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "payTo": {
            "type": "string",
            "description": "Receive address (0x… for EVM, base58 for Solana)."
          },
          "maxTimeoutSeconds": {
            "type": "integer"
          },
          "asset": {
            "type": "string",
            "description": "USDC contract address or SPL mint."
          },
          "outputSchema": {
            "type": "object"
          },
          "extra": {
            "type": "object",
            "description": "EVM: EIP-712 domain (`name`, `version`). Solana: `feePayer`, the sponsor that pays the SOL network fee."
          }
        }
      },
      "PaymentRequired": {
        "type": "object",
        "required": [
          "x402Version",
          "accepts"
        ],
        "properties": {
          "x402Version": {
            "type": "integer",
            "enum": [
              1
            ]
          },
          "error": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "accepts": {
            "type": "array",
            "description": "Both rails: USDC on Base and USDC on Solana.",
            "items": {
              "$ref": "#/components/schemas/PaymentRequirements"
            }
          }
        }
      }
    }
  }
}